requirements.txt
pandas>=1.3.4
scikit-learn>=0.24.2
flask>=1.1.2
tensorflow>=2.7.0
pillow>=8.4.0
numpy>=1.20.3
keras>=2.7.0
gunicorn>=19.9.0
